+++ include/tfrt/tensor/tensor_shape.h
@@ -187,12 +187,12 @@ class PartialTensorShape {
   // Create a PartialTensorShape with the dimensions. If rank itself is unknown,
   // (dims is llvm::None), this is an unranked. Else, it is ranked where each
   // dimensions could still be unknown (indicated by kUnknownDimSize) as well.
-  explicit PartialTensorShape(Optional<ArrayRef<int64_t>> dims);
+  explicit PartialTensorShape(Optional<ArrayRef<ssize_t>> dims);
 
   // Returns the shape of the tensor.
   // If unranked, return llvm::None
   // If ranked, return dimensions (including kUnknownDimSize for unknown dim).
-  Optional<ArrayRef<int64_t>> GetShape() const;
+  Optional<ArrayRef<ssize_t>> GetShape() const;
 
   // Returns true if the rank is unknown. Else, returns false.
   // TODO(ashwinm): Add a test for unranked in tensor_shape.mlir
@@ -218,7 +218,7 @@ class PartialTensorShape {
   // We store dims in SmallVector here since PartialTensorShape is designed
   // for use in shape computations where we could alter the shape by adding/
   // removing dimensions.
-  Optional<SmallVector<int64_t, 4>> dims_;
+  Optional<SmallVector<ssize_t, 4>> dims_;
 };
 
 //
